pyproject.toml
[project]
name = "aws-gpt-oss-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.11",
    "openai>=1.99.6",
    "python-dotenv>=1.1.1",
    "strands-agents[openai]>=1.4.0",
]
